Light tweaks. Played around with doing 4 bit color channels for lighting but it ran really slow or I am an idiot. There are still bugs in the lighting algorithm but so far I am faking it.

+++ src/jamTiles.h
@@ -4,9 +4,9 @@
 #include "jamTypes.h"
 
 struct tile {
-  i32 type;
+  s32 type;
   // shader information
-  u8 light;
+  u16 light;
 };
 
 struct world {
